+++ cpp-terminal/iostream.cpp
@@ -1,6 +1,4 @@
 #include "cpp-terminal/iostream.hpp"
-
-#include "cpp-terminal/io.hpp"
 #include "cpp-terminal/stream.hpp"
 
 #include <new>
