+++ cpp-terminal/terminal.cpp
@@ -7,6 +7,33 @@
 #include "cpp-terminal/screen.hpp"
 #include "cpp-terminal/style.hpp"
 
+
+#include <new>
+
+namespace Term
+{
+static char termbuf[sizeof(Term::Terminal)];
+Terminal&   terminal = reinterpret_cast<Term::Terminal&>(termbuf);
+}  // namespace Term */
+
+int Term::TerminalInitializer::m_counter{0};
+
+void Term::TerminalInitializer::init()
+{
+  if(m_counter++ == 0) new(&Term::terminal) Terminal();
+}
+
+Term::TerminalInitializer::TerminalInitializer()
+{
+  init();
+}
+
+Term::TerminalInitializer::~TerminalInitializer()
+{
+  if(--m_counter == 0) (&Term::terminal)->~Terminal();
+}
+
+
 void Term::Terminal::setOptions(const Term::Options& options)
 {
   m_options = options;
