+++ examples/minimal.cpp
@@ -1,4 +1,4 @@
-#include "cpp-terminal/io.hpp"
+#include "cpp-terminal/terminal.hpp"
 
 #include <iostream>
