+++ tests/attach_console.test.cpp
@@ -1,6 +1,5 @@
 #include "cpp-terminal/color.hpp"
 #include "cpp-terminal/exception.hpp"
-#include "cpp-terminal/io.hpp"
 #include "cpp-terminal/style.hpp"
 #include "cpp-terminal/tty.hpp"
 #include "cpp-terminal/version.hpp"
